Gameswitcher save/load feature as a tweak
