+++ src/gameSwitcher/res/gs-legend-v2.png
