--- src/gameSwitcherMod/Makefile
@@ -1,8 +1,0 @@
-INCLUDE_CJSON=1
-include ../common/config.mk
-
-TARGET = gameSwitcherMod
-LDFLAGS := $(LDFLAGS) -lSDL -lSDL_ttf -lSDL_image -lSDL_mixer -lSDL_rotozoom -lpng -lsqlite3 -pthread
-
-include ../common/commands.mk
-include ../common/recipes.mk
