--- src/gameSwitcherModInstaller/Makefile
@@ -1,8 +1,0 @@
-INCLUDE_CJSON=1
-include ../common/config.mk
-
-TARGET = gameSwitcherModInstaller
-LDFLAGS := $(LDFLAGS) -lSDL -lSDL_image -lSDL_ttf -lSDL_mixer -lSDL_rotozoom -lpng
-
-include ../common/commands.mk
-include ../common/recipes.mk
