+++ src/tweaks/appstate.h
@@ -68,6 +68,7 @@ static List _menu_advanced;
 static List _menu_reset_settings;
 static List _menu_tools;
 static List _menu_diagnostics;
+static List _menu_gameswitcher;
 
 void menu_free_all(void)
 {
@@ -86,6 +87,7 @@ void menu_free_all(void)
     list_free(&_menu_reset_settings);
     list_free(&_menu_tools);
     list_free(&_menu_diagnostics);
+    list_free(&_menu_gameswitcher);
 
     menu_icons_free_all();
     menu_network_free_all();
