--- static/packages/App/GameSwitcher Mod (with save n load)/App/GameSwitcherMod/config.json
@@ -1,6 +1,0 @@
-{
-	"label":	"GameSwitcher Mod",
-	"icon":	"/mnt/SDCARD/Icons/Default/app/gameswitcher.png",
-	"launch":	"launch.sh",
-	"description":	"Quickly switch games with save & load"
-}
