--- static/packages/App/GameSwitcher Mod (with save n load)/App/GameSwitcherMod/launch.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-echo $0 $*
-cd $(dirname "$0")
-./gameSwitcherModInstaller
